requirements.txt
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
requests>=2.28.2
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
bitsandbytes>=0.41.0
safetensors>=0.3.1
psutil>=5.9.0
GPUtil>=1.4.0